Update form input & add class to results p tag
1. Update form input
- Add oninput attribute to prevent the input of
either negative numbers or non-number values
- Floating point numbers will also be prevented by
removing the decimal place upon input
For example: 1.2 becomes 12

2. Add class to results p tag
- Since a p tag is being used to represent the
fizz buzz results, a placeholder attribute is not
available. To get around this, initial text is
added, and a before-results class is added and
styled accordingly. Javascript will be added to
change the class and update the styling
accordingly.